+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/mapping/service/impl/PollingFederationService.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2019 Expedia, Inc.
+ * Copyright (C) 2016-2022 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,17 +21,29 @@ import java.util.Map;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import io.micrometer.core.instrument.Counter;
+import io.micrometer.core.instrument.ImmutableTag;
+import io.micrometer.core.instrument.MeterRegistry;
+
+import com.google.common.annotations.VisibleForTesting;
 
 import com.hotels.bdp.waggledance.api.model.AbstractMetaStore;
 import com.hotels.bdp.waggledance.api.model.MetaStoreStatus;
 import com.hotels.bdp.waggledance.core.federation.service.PopulateStatusFederationService;
 
+@Service
 public class PollingFederationService {
 
   private final static Logger log = LoggerFactory.getLogger(PollingFederationService.class);
+  private final static String METASTORE_STATUS_METRIC_NAME = "metastore_status";
+  private final static String METASTORE_TAG_NAME = "metastore";
 
   private final PopulateStatusFederationService populateStatusFederationService;
   private Map<String, MetaStoreStatus> previous = new HashMap<>();
+  private @Autowired MeterRegistry meterRegistry;
 
   public PollingFederationService(PopulateStatusFederationService populateStatusFederationService) {
     this.populateStatusFederationService = populateStatusFederationService;
@@ -42,7 +54,10 @@ public class PollingFederationService {
     Map<String, MetaStoreStatus> current = new HashMap<>();
     List<AbstractMetaStore> metastores = populateStatusFederationService.getAll();
     for (AbstractMetaStore metaStore : metastores) {
-      current.put(metaStore.getName(), metaStore.getStatus());
+      String metastoreName = metaStore.getName();
+      MetaStoreStatus metastoreStatus = metaStore.getStatus();
+      current.put(metastoreName, metastoreStatus);
+      sendMetric(metastoreName, metastoreStatus);
       MetaStoreStatus previousMetastoreStatus = previous.get(metaStore.getName());
       if (previousMetastoreStatus != null) {
         if (previousMetastoreStatus != metaStore.getStatus()) {
@@ -51,5 +66,18 @@ public class PollingFederationService {
       }
     }
     previous = current;
+  }
+
+  private void sendMetric(String metastoreName, MetaStoreStatus status) {
+    ImmutableTag tag = new ImmutableTag(METASTORE_TAG_NAME, metastoreName);
+    Counter counter = Counter.builder(METASTORE_STATUS_METRIC_NAME)
+        .tag(tag.getKey(), tag.getValue())
+        .register(meterRegistry);
+    counter.increment(status.getIntValue());
+  }
+
+  @VisibleForTesting
+  void setMeterRegistry(MeterRegistry meterRegistry) {
+    this.meterRegistry = meterRegistry;
   }
 }