+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+
 ## [3.10.5] - 2022-05-23
 ### Changed
 * Added `queryFunctionsAcrossAllMetastores` configuration for optimising `getAllFunctions` calls.
+### Added
+* Metrics to track metastore availability. 
 
 ## [3.10.4] - 2022-04-17
 ### Fixes